Create maverick-master-pihole-blocklist
This is a list of the blocklist I have compiled with hours of scouring the internet. It blocks drugs, crpyto, tracking, ads, malware, ransomware, some porn. Super smexy. It still has some work to be done on a few out of the many lists to make it readable by pihole. Help me out and branch off with corrections or ask me and I'll update the list URLs format properly so pihole can read it!